--- l10n_br_purchase/res_company_view.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-	<data>
-
-		<record id="view_l10n_br_purchase_company_form" model="ir.ui.view">
-		<field name="name">l10n_br_purchase.res.company.form</field>
-		<field name="model">res.company</field>
-		<field name="inherit_id" ref="l10n_br_account_product.view_l10n_br_account_company_form" />
-			<field name="arch" type="xml">
-				<page position="before" string="NFe">
-					<page string="Compras">
-						<group string="Configurações de Compras">
-							<field name="purchase_fiscal_category_id" />
-						</group>
-					</page>
-				</page>
-			</field>
-		</record>
-
-	</data>
-</openerp>
